# Conflicts:
#	iOS/Med2Med/Supporting Files/Med2Med-Info.plist

+++ iOS/Shared/Models/AuthenticationModel.m
@@ -184,7 +184,7 @@
 		{
 			NSLog(@"Retry Refreshing Tokens");
 			
-			dispatch_after(dispatch_time(DISPATCH_TIME_NOW, 2.0 * NSEC_PER_SEC), dispatch_get_main_queue(), ^
+			dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (int64_t)(2.0 * NSEC_PER_SEC)), dispatch_get_main_queue(), ^
 			{
 				[self getNewTokensWithSuccess:success failure:failure isRetry:YES];
 			});
